refactor: move footer components into a footer folder
